requirements.txt
python-dotenv
fredapi
tastytrade
scipy
numba
